requirements.txt
PyYAML
boto3
simplejson
